+++ example/lib/main.dart
@@ -1,44 +1,41 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_multiselect/flutter_multiselect.dart';
 
+import 'cities.dart';
+
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.blue,
-        //brightness: Brightness.dark,
-      ),
-      home: MyHomePage(title: 'Flutter Demo - Multiselect'),
-    );
+    return ValueListenableBuilder(
+        valueListenable: _rebuilder,
+        builder: (context, int value, child) {
+          return MaterialApp(
+            key: ValueKey(value),
+            title: 'Flutter Demo',
+            debugShowCheckedModeBanner: false,
+            themeMode: value.isOdd ? ThemeMode.light : ThemeMode.dark,
+            theme: ThemeData(primarySwatch: Colors.blue, brightness: Brightness.light),
+            darkTheme: ThemeData(primarySwatch: Colors.red, brightness: Brightness.dark),
+            home: MyHomePage(title: 'Flutter Demo - Multiselect'),
+          );
+        });
   }
 }
 
+final _RebuildApp _rebuilder = _RebuildApp();
+
+void rebuildApp() => _rebuilder.execute();
+
+class _RebuildApp extends ValueNotifier<int> {
+  _RebuildApp() : super(1);
+
+  void execute() => value = value + 1;
+}
+
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
+  MyHomePage({Key? key, this.title = ''}) : super(key: key);
 
   final String title;
 
@@ -47,54 +44,76 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
+        actions: [
+          ElevatedButton(
+              onPressed: () => rebuildApp(),
+              child: _rebuilder.value.isOdd ? Text('Switch to Dark theme') : Text('Switch to Light theme'))
+        ],
       ),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: MultiSelect(
-                autovalidate: true,
-                maxLength: 5, // optional
-                validator: (dynamic value) {
-                  if (value == null) {
-                    return 'Please select one or more option(s)';
-                  }
-                  return null;
-                },
-                errorText: 'Please select one or more option(s)',
-                dataSource: [
-                  {"name": "Afghanistan", "code": "AF"},
-                  {"name": "Åland Islands", "code": "AX"},
-                  {"name": "Albania", "code": "AL"},
-                ],
-                textField: 'name',
-                valueField: 'code',
-                filterable: true,
-                required: true,
-                onSaved: (value) {
-                  print('The selected values are $value');
-                },
-                //selectIcon: Icons.arrow_drop_down_circle,
-                //saveButtonColor: Theme.of(context).primaryColor,
-                //checkBoxColor: Theme.of(context).primaryColorDark,
-                //cancelButtonColor: Theme.of(context).primaryColorLight,
+        child: Form(
+          key: _formKey,
+          autovalidateMode: AutovalidateMode.always,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: MultiSelect(
+                  autovalidateMode: AutovalidateMode.always,
+                  initialValue: ['IN', 'US'],
+                  titleText: 'Country of Residence',
+                  maxLength: 5, // optional
+                  validator: (dynamic value) {
+                    return value == null ? 'Please select one or more option(s)' : null;
+                  },
+                  errorText: 'Please select one or more option(s)',
+                  dataSource: cities,
+                  textField: 'name',
+                  valueField: 'code',
+                  filterable: true,
+                  required: true,
+                  onSaved: (value) {
+                    print('The saved values are $value');
+                  },
+                  change: (value) {
+                    print('The selected values are $value');
+                  },
+                  selectIcon: Icons.arrow_drop_down_circle,
+                  saveButtonColor: Theme.of(context).primaryColor,
+                  checkBoxColor: Theme.of(context).primaryColorDark,
+                  cancelButtonColor: Theme.of(context).primaryColorLight,
+                  responsiveDialogSize: Size(600, 800),
+                ),
               ),
-            ),
+              SizedBox(width: 10.0),
+              ElevatedButton(
+                child: Text('Save'),
+                style: ElevatedButton.styleFrom(
+                  primary: Theme.of(context).colorScheme.primary,
+                  onPrimary: Theme.of(context).colorScheme.onPrimary,
+                ),
+                onPressed: () {
+                  _onFormSaved();
+                }),
             SizedBox(
               width: 10.0,
             ),
           ],
         ),
       ),
-
-      // This trailing comma makes auto-formatting nicer for build methods.
+    )
     );
+  }
+
+  void _onFormSaved() {
+    final FormState? form = _formKey.currentState;
+    form?.save();
   }
 }