+++ example_customization_multiselect/lib/main.dart
@@ -29,7 +29,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title = ''}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -67,7 +67,7 @@ class _MyHomePageState extends State<MyHomePage> {
                 checkBoxColor: Colors.black,
                 selectedOptionsInfoText: "Selected custom text (tap to remove)",
                 selectedOptionsBoxColor: Colors.green,
-                autovalidate: true,
+                autovalidateMode: AutovalidateMode.always,
                 maxLength: 5, // optional
                 //--------end customization selection modal------------
                 validator: (dynamic value) {